fixed issues with image and code readability.
